+++ src/wasm/openscad.fonts.d.ts
@@ -1,0 +1,2 @@
+import OpenSCAD from "./openscad.wasm";
+export declare function addFonts(openscad: OpenSCAD.Instance): void;